+++ src/Checking.java
@@ -1,5 +1,5 @@
 public class Checking extends Account {
-    public Checking(Id accountId, Password accountPassword, int serviceFee) {
+    public Checking(Id accountId, Password accountPassword, double serviceFee) {
         super(accountId, accountPassword, serviceFee, "Checking");
     }
 }
